feat: Merge the tuner_v3 with libnccl-net.so
This change links the nccl_ofi_tuner v3_api with the plugin.
This will ensure that the tuner is loaded automatically when
the nccl_ofi_plugin is loaded. This also adds a new envirnment
variable OFI_NCCL_ENABLE_TUNER to be set to use the tuner
internally with the plugin.