scheduler: Change the way the num_of_stripes is calculated
Currently no_stripes is calulated with the below formula :
num_stripes = MIN(max_nr_rails, CEIL(payload_size / min_stripe_size))
This can result in number of stripes being 3 on a 4 rail system.
This commit modifies the  num_stripes to be be  rounded off
to the nereast factor of the num_rails.

Signed-off-by: Arun Karthik <akkart@amazon.com>